+++ packages/contracts/test/voting/admin-setup.ts
@@ -43,7 +43,6 @@ describe('AdminSetup', function () {
 
     const iface = new ethers.utils.Interface([
       'function initialize(address  _dao)',
-      'function proposalCount()',
       'function executeProposal(bytes _metadata, tuple(address,uint256,bytes)[] _actions)',
     ]);