+++ packages/subgraph/src/utils/constants.ts
@@ -14,7 +14,7 @@ export const VOTING_MODES = new Map<number, string>()
 
 export const TOKEN_VOTING_INTERFACE = '0x302aac6a';
 export const ADDRESSLIST_VOTING_INTERFACE = '0xcedebd0f';
-export const ADMIN_INTERFACE = '0x4b37035a';
-export const MULTISIG_INTERFACE = '0xcc7e98bb';
+export const ADMIN_INTERFACE = '0x9102c53e';
+export const MULTISIG_INTERFACE = '0x164b5edf';
 
 export const TEN_POWER_16 = '10000000000000000';